NumEx v0.1.0 Binary

A module to perform binary operations on enumerables

Functions

binary_repr(x)

Takes an Integer, returns its binary representation.

Examples

iex>Binary.binary_repr(5)
101

bitwise_and(arr1, arr2)

Takes an two enumerables, performs bitwise_and on them, element_wise.

Examples

iex> Binary.bitwise_and([2,3,4],[4,5,6])
[0, 1, 4]

bitwise_or(arr1, arr2)

Takes an two enumerables, performs bitwise_or on them, element_wise.

Examples

iex> Binary.bitwise_or([2,3,4],[4,5,6])
[6, 7, 6]

bitwise_xor(arr1, arr2)

Takes an two enumerables, performs bitwise_or on them, element_wise.

Examples

iex> Binary.bitwise_xor([2,3,4],[4,5,6])
[6, 6, 2]

invert(arr1)

Takes an enumerable, performs bitwise_not on its elements.

Examples

iex>  Binary.invert([32, 5, 1])
[-33, -6, -2]

left_shift(x, y)

Takes two values, performs arithmetic left shift on them.

Examples

iex> Binary.left_shift(2,[4,5,6])
[32, 64, 128]

right_shift(x, y)

Takes two values, performs arithmetic right shift on them.

Examples

iex> Binary.right_shift(32,[4,5,6])
[2, 1, 0]