NumEx v0.1.0 MathematicalFunc

A module to perform mathematical unctions

Functions

acos(list)

Takes a list of elements and computes the acos function element wise .

iex>MathematicalFunc.acos([1,2,34,4,5])

acosh(list)

Takes a list of elements and computes the acosh function element wise .

iex>MathematicalFunc.acosh([1,2,34,4,5])

asin(list)

Takes a list of elements and computes the asin function element wise .

iex>MathematicalFunc.asin([1,2,34,4,5])

asinh(list)

Takes a list of elements and computes the asin function element wise .

iex>MathematicalFunc.asin([1,2,34,4,5])

atan(list)

Takes a list of elements and computes the atan function element wise .

iex>MathematicalFunc.atan([1,2,34,4,5])

atanh(list)

Takes a list of elements and computes the atanh function element wise .

iex>MathematicalFunc.atanh([1,2,34,4,5])

ceil(l)

cos(list)

Takes a list of elements and computes the cos function element wise .

iex>MathematicalFunc.cos([1,2,34,4,5])

cosh(list)

Takes a list of elements and computes the cosh function element wise .

iex>MathematicalFunc.cosh([1,2,34,4,5])

deg2rad(degree)

Convets the angle value in degrees to radians .

iex> MathematicalFunc.deg2rad(180)
3.14159

exmp(l)

exp(l)

exp2(l)

floor(l)

hypot(s1, s2)

Takes the values of two sides of a right angled triangle and gives the hypotenuse .

iex> MathematicalFunc.hypot(3, 4)
5.0

rad2deg(radian)

Convets the angle value in degrees to radians .

iex> MathematicalFunc.rad2deg(3.14159)
180

rint(l)

sin(list)

Takes a list of elements and computes the sine function element wise .

##Examples

iex>MathematicalFunc.sin([1,2,34,4,5])

sinh(list)

Takes a list of elements and computes the Hyperbolic sine function element wise .

iex>MathematicalFunc.sinh([1,2,34,4,5])

tan(list)

Takes a list of elements and computes the tan function element wise .

iex>MathematicalFunc.tan([1,2,34,4,5])

tanh(list)

Takes a list of elements and computes the tanh function element wise .

iex>MathematicalFunc.tanh([1,2,34,4,5])

truncate(l)