NumEx v0.1.0 StringManipulations

A module to perform operations on Strings.

Functions

add(s1, s2)

add(String.t(), String.t()) :: list()

Takes two strings and concatenates them element-wise.

Examples

iex> StringManipulations.add("Elixir", "Lang")
["EL", "la", "in", "xg"]

center(s1, n)

center(String.t(), integer()) :: String.t()

Return a copy of s1 with its elements centered in a string of length width..

Examples

iex> StringManipulations.center("Elixir", 5)
"Elixi"

join(s1, seq)

join(String.t(), String.t()) :: String.t()

Return a string which is the concatenation of the strings in the sequence seq.

Examples

iex> StringManipulations.join("Elixir", "Lang")
"LElixiraElixirnElixirg"

multiply(s1, n)

multiply(String.t(), integer()) :: list()

Return (a * i), that is string multiple concatenation, element-wise.

Examples

iex> StringManipulations.multiply("Elixir", 3)
["EEE", "lll", "iii", "xxx", "iii", "rrr"]