NumEx v0.1.0 MathematicalFunc

A module to perform mathematical functions

Functions

acos(list)

Takes a list of elements and computes the acos function element wise .

Examples

iex>MathematicalFunc.acos([1,2,34,4,5])

acosh(list)

Takes a list of elements and computes the acosh function element wise .

Examples

iex>MathematicalFunc.acosh([1,2,34,4,5])

add(list)

Add arguments element-wise.

Examples

iex> MathematicalFunc.add([4, 6, 8190])
8200

asin(list)

Takes a list of elements and computes the asin function element wise .

Examples

iex>MathematicalFunc.asin([1,2,34,4,5])

asinh(list)

Takes a list of elements and computes the asin function element wise .

Examples

iex>MathematicalFunc.asin([1,2,34,4,5])

atan(list)

Takes a list of elements and computes the atan function element wise .

Examples

iex>MathematicalFunc.atan([1,2,34,4,5])

atanh(list)

Takes a list of elements and computes the atanh function element wise .

Examples

iex>MathematicalFunc.atanh([1,2,34,4,5])

ceil(list)

Return the ceiling of the input, element-wise.

Examples

iex> MathematicalFunc.ceil([3.89, 4.70, 5.62])
[4.0, 5.0, 6.0]

cos(list)

Takes a list of elements and computes the cos function element wise .

Examples

iex>MathematicalFunc.cos([1,2,34,4,5])

cosh(list)

Takes a list of elements and computes the cosh function element wise .

Examples

iex>MathematicalFunc.cosh([1,2,34,4,5])

deg2rad(degree)

Convets the angle value in degrees to radians .

Examples

iex> MathematicalFunc.deg2rad(180)
3.14159

exmp(list)

Calculate exp(x) - 1 for all elements in the array.

Examples

iex> MathematicalFunc.exmp([2, 3, 5])
[6.38, 19.08, 147.41]

exp(list)

Calculate the exponential of all elements in the input array.

Examples

iex> MathematicalFunc.exp([2,3,5])
[7.38, 20.08, 148.41]

exp2(list)

Calculate :math.pow(2, x) for all x in the input array.

Examples

iex> MathematicalFunc.exp2([4, 5, 6])
[16.0, 32.0, 64.0]

floor(list)

Return the floor of the input, element-wise.

Examples

iex> MathematicalFunc.floor([3.89, 4.70, 5.62])
[3.0, 4.0, 5.0]

hypot(s1, s2)

Takes the values of two sides of a right angled triangle and gives the hypotenuse .

Examples

iex> MathematicalFunc.hypot(3, 4)
5.0

log(list)

Natural logarithm, element-wise.

Examples

iex> MathematicalFunc.log([3, 4, 5, 8])
[0.47, 0.60, 0.69, 0.90]

log10(list)

Return the base 10 logarithm of the input array, element-wise.

Examples

iex> MathematicalFunc.log10([10, 100, 1000])
[1.0, 2.0, 3.0]

log1p(list)

Return the natural logarithm of one plus the input array, element-wise.

Examples

iex> MathematicalFunc.log1p([4, 5, 7, 8])
  [1.61, 1.79, 2.07, 2.19]

log2(list)

Base-2 logarithm of x for all x in input array.

Examples

iex> MathematicalFunc.log2([2,4, 8])
[1.0, 2.0, 3.0]

logaddexp(x1, x2)

Logarithm of the sum of exponentiations of the inputs.

Examples

iex> MathematicalFunc.logaddexp(5, 10)
10.006

logaddexp2(x1, x2)

Logarithm of the sum of exponentiations of the inputs in base-2.

Examples

iex> MathematicalFunc.logaddexp2(2, 3)
2.48

rad2deg(radian)

Convets the angle value in radians to degrees.

Examples

iex> MathematicalFunc.rad2deg(3.14159)
180

rint(list)

Round elements of the array to the nearest integer.

Examples

iex> MathematicalFunc.rint([3.89, 4.70, 5.62])
[4, 5, 6]

sin(list)

Takes a list of elements and computes the sine function element wise .

##Examples

iex>MathematicalFunc.sin([1,2,34,4,5])
["0.84", "0.91", "0.53", "-0.76", "-0.96"]

sinc(list)

Return the sinc function.

Examples

iex> MathematicalFunc.sinc([4, 6, -4, 5])
[-0.09, -0.03, -0.09, 0.06]

sinh(list)

Takes a list of elements and computes the Hyperbolic sine function element wise .

Examples

iex>MathematicalFunc.sinh([1,2,34,4,5])

tan(list)

Takes a list of elements and computes the tan function element wise .

Examples

iex>MathematicalFunc.tan([1,2,34,4,5])

tanh(list)

Takes a list of elements and computes the tanh function element wise .

Examples

iex>MathematicalFunc.tanh([1,2,34,4,5])

truncate(list)

Return the truncated value of the input, element-wise.

Examples

iex> MathematicalFunc.truncate([3.89, 4.70, 5.62])
[3, 4, 5]