NumEx v0.1.0 Matrix

A module to perform Matrix Operations

Functions

add(m1, m2)

add([[number()]], [[number()]]) :: [[number()]]

Performs elmentwise addition

Examples

iex> Matrix.add([[1, 2, 3],[1, 1, 1]],[[3, 2, 1],[2, 2, 2]])
[[4, 4, 4],
 [3, 3, 3]]

columncount(list)

columncount([[number()]]) :: number()

Takes matrix and returns number of columns in the given matrix.

Examples

iex> Matrix.colcount([[12,34,56,11],[98,78,67,0],[0,34,78,9]])
4

eye(row, col, k)

eye(number(), number(), number()) :: [[number()]]

Takes three numerals ‘row’, ‘col’, ‘k’ and returns eye matrix of the given order.

Examples

iex> Matrix.eye(4,3,0)
[[1, 0, 0], [0, 1, 0], [0, 0, 1], [0, 0, 0]]
iex> Matrix.eye(3, 3, -1)
[[0, 0, 0], [1, 0, 0], [0, 1, 0]]

genrandrow(col)

genrandrow(number()) :: [number()]

Takes a numeral ‘col’ and returns a single row of ‘col’ number of random elements..

Examples

iex> Matrix.genrandrow(6)
[34, 101, 23, 110, 17, 10]

identity(row, col)

identity(number(), number()) :: [[number()]]

Takes two numerals ‘row’, ‘col and returns a the identity matrix of the given order.

Examples

iex> Matrix.identity(6,3)
[[1, 0, 0], [0, 1, 0], [0, 0, 1], [0, 0, 0], [0, 0, 0], [0, 0, 0]]

max(mat)

max([[number()]]) :: [number()]

Takes a Matrix and finds the maximum element from each row.

Examples

iex>Matrix.max([[1,2,34],[3,4,5],[4,5,6,7]])
       [34, 5, 7]

ones(row, col)

ones(number(), number()) :: [[number()]]

Takes two numerals ‘row’, ‘col’ and returns a matrix of ones.

Examples

iex> Matrix.ones(5,3)
[[1, 1, 1], [1, 1, 1], [1, 1, 1], [1, 1, 1], [1, 1, 1]]

rand(row, col)

rand(number(), number()) :: [[number()]]

Takes two numerals ‘row’, ‘col’ and returns a matrix of random elements.

Examples

iex> Matrix.rand(2,3)
[[47, 31, 53], [6, 7, 42]]

rowcount(row)

rowcount([[number()]]) :: number()

Takes matrix and returns number of rows in the given matrix.

Examples

iex> Matrix.rowcount([[12,34,56,11],[98,78,67,0],[0,34,78,9]])
3

scalar(m1, s)

scalar([[number()]], number()) :: [[number()]]

Elementwise mutiplication with a scalar.

Examples

iex> Matrix.scalar([[2, 2, 2], [1, 1, 1]], 2)
[[4, 4, 4],
 [2, 2, 2]]

size(row)

size([number()]) :: number()

Takes a Matrix, returns size of the matrix.

Examples

iex> Matrix.size([[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0]])
     6

sub(m1, m2)

sub([[number()]], [[number()]]) :: [[number()]]

Performs elementwise subtraction

Examples

iex> Matrix.sub([[1, 2, 3],[1, 2, 2]],[[1, 2, 3],[2, 2, 2]])
[[0, 0, 0],
 [-1, 0, 0]]

transpose(a)

transpose([[number()]]) :: [[number()]]

Transposes a matrix.

Examples

iex> Matrix.transpose([[1, 2, 3], [4, 5, 6]])
[[1, 4], [2, 5], [3, 6]]

zeros(row, col)

zeros(number(), number()) :: [[number()]]

Takes two numerals ‘row’, ‘col’ and returns a matrix of zeros.

Examples

iex> Matrix.zeros(6,3)
[[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0]]