NumEx v0.1.0 Sort

A module to sort enumerables.

Functions

lexsort(list1, list2)

Takes a two enumerables and performs Lexographic sort operations.

iex> Sort.lexsort([5, 3, 6, 8, 1], [0, 9, 4, 7, 2]) [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]

msort(list)

Return a sorted copy of an Enumerable sorted along one axis only

iex> Sort.msort([[3, 4], [0, 1], [6, 5]])
  [[0, 1], [3, 4], [5, 6]]

quicksort(list)

Takes a list of elements and sorts the elements in quick sort fashion.

iex>  Sort.quicksort([195,99,78,12])
 [12, 78, 99, 195]]

sort(list, axis)

Return a sorted copy of a List of Lists based on the value of axis

iex> Sort.sort([[6, 2], [5, 7], [0, 9]], 1) [[2, 6], [5, 7], [0, 9]] iex> Sort.sort([[6, 2], [5, 7], [1, 3]], 0) [[1, 2], [3, 5], [6, 7]]